# faster ScmRun.__setitem__