# Pull request

Please confirm that this pull request has done the following:

- [x] Tests added
- [x] Documentation added (where applicable)
- [ ] Example added (either to an existing notebook or as a new notebook, where applicable)
- [x] Description in ``CHANGELOG.rst`` added

# Content

Create the changed MultiIndex via levels and codes instead of via a DataFrame for better performance.

# Benchmarks

I benchmarked setting metadata on a run with 5000 rows. I only benchmarked overwriting existing metadata for practical reasons, but the performance should be similar.

<details><summary>Benchmarking code</summary>

```python

import scmdata
import numpy as np


def create_run():
    years = np.arange(1990, 2010)
    n = 5000
    return scmdata.ScmRun(
        np.ones((len(years), n), dtype=float),
        years,
        {
            "variable": [str(x) for x in range(n)],
            "unit": ["Mt CO2 / year"] * n,
            "category": ["0"] * n,
            "model": ["TEST"] * n,
            "scenario": ["TEST"] * n,
            "region": ["WORLD"] * n,
            "stage": ["test"] * n,
        },
    )


import timeit

single_existing = timeit.timeit(
    "a['scenario'] = 'This is not a test'",
    setup="a = create_run()",
    globals=globals(),
    number=1000,
)
print(f"{single_existing=}")
multi_existing = timeit.timeit(
    "a['scenario'] = b",
    setup="a = create_run(); b = ['This is not a test'] * 5000",
    globals=globals(),
    number=1000,
)
print(f"{multi_existing=}")
```
</details>


old, 1000 executions:
single_existing=1.4 s
multi_existing=1.3 s

new, 1000 executions:
single_existing=0.20 s
multi_existing=0.79 s


So, for the common case of replacing all values with a single new value, we get a 7 times speedup, for the case where we replace with a full set of new values, the speedup is 1.6.