@mikapfl 2023-06-29
In the ndc-quantification test suite, I see a 20% speed-up (1:55 min instead of 2:19 min) with this fix, which is quite nice given that it is a micro-optimization which shows measurable gains in a macro context.